+++ src/sn/FocusMng.ts
@@ -21,7 +21,7 @@ export class FocusMng {
 	#idx					= -1;
 	readonly	#elc		= new EventListenerCtn;
 
-	destroy() {this.#aBtn = []; this.#idx = -1; this.#elc.clear();}
+	destroy() {this.#aBtn = []; this.#idx = -1; this.#elc.clear()}
 
 	add(cmp: Container | HTMLElement, on: ()=> boolean, off: ()=> void) {
 //console.log(`fn:FocusMng.ts line:62 ADD idx:${this.aBtn.length} local:${(cmp as HTMLElement).localName} type:${(cmp as HTMLInputElement).type} cmp:%o`, cmp);
@@ -159,7 +159,7 @@ export class FocusMng {
 		return b.on() ?b.btn : null;
 	}
 
-	blur() {this.#allOff(); this.#idx = -1; globalThis.focus();}
+	blur() {this.#allOff(); this.#idx = -1; globalThis.focus()}
 	#allOff() {
 		for (let i=this.#aBtn.length -1; i>=0; --i) {
 			const b = this.#aBtn[i];